+++ downloader/downloader.ts
@@ -49,11 +49,13 @@ export class Downloader {
         const segments: Array<string> = fs.readdirSync(segmentsDir).map(it => segmentsDir + "/" + it);
         segments.sort();
 
+        if (segments.length == 0) return
+
         // Merge TS files
         await mergeFiles(segments, mergedSegmentsFile);
 
         // Transmux
-        await transmuxTsToMp4(mergedSegmentsFile, segmentsDir + "video.mp4");
+        await transmuxTsToMp4(mergedSegmentsFile, segmentsDir + "/video.mp4");
 
         // Delete ts files
         fs.remove(mergedSegmentsFile);
